requirements.txt
numpy
pandas
pydeck
google-generativeai
python-dotenv
matplotlib
seaborn
anthropic
flask
openai
highcharts-core
mistralai
python-pptx
pillow
functions
cohere
azure-ai-inference
azure-core